% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{add_logo}
\alias{add_logo}
\title{Add CORI logo to an image}
\usage{
add_logo(
  plot_path,
  logo_path = "https://rwjf-public.s3.amazonaws.com/Logo-Mark_CORI_Black.svg",
  logo_position = "top right",
  logo_scale = 16
)
}
\arguments{
\item{plot_path}{file path to the chart/plot/graphic}

\item{logo_path}{Path to the logo. Defaults to hosted Full CORI Black logo}

\item{logo_position}{Combination of top/bottom and right/left. Defaults to top right.}

\item{logo_scale}{Scale logo to 1/10 width of plot}
}
\value{
magick image
}
\description{
Add CORI logo to an image
}
