% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadband.R
\name{add_scale_and_north_arrow}
\alias{add_scale_and_north_arrow}
\title{Add a scale and north arrow to a map}
\usage{
add_scale_and_north_arrow(fig, data, font_family = "Bitter")
}
\arguments{
\item{fig}{ggplot map}

\item{data}{spatial data to use for drawing the scale}

\item{font_family}{font family to use for the scale label}
}
\value{
Updated ggplot figure
}
\description{
Add a scale and north arrow to a map
}
