% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cori_communities}
\alias{cori_communities}
\title{CORI network communities}
\value{
\item{cori_communities}{a data frame}
}
\description{
Communities that are part of the Center for Rural Innovation network
}
\examples{
data(cori_communities)
str(cori_communities)
}
\keyword{datasets}
