% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cori_education}
\alias{cori_education}
\title{Working remotely by education}
\value{
\item{cori_education}{a data frame}
}
\description{
Time series data on what percent of population are working remotely by
education
}
\examples{
data(cori_education)
str(cori_education)
}
\keyword{datasets}
