% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cori_employment}
\alias{cori_employment}
\title{Employment in CORI network communities}
\value{
\item{cori_employment}{a data frame}
}
\description{
Estimates for total population and employment in 2014 and 2019 for ten
CORI communities
}
\examples{
data(cori_employment)
str(cori_employment)
}
\keyword{datasets}
