% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cori_poverty}
\alias{cori_poverty}
\title{Poverty in CORI network communities}
\value{
\item{cori_poverty}{a data frame}
}
\description{
Estimates for total population and population below the poverty level in
2014 and 2019 for ten CORI communities
}
\examples{
data(cori_poverty)
str(cori_poverty)
}
\keyword{datasets}
