% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/broadband.R
\name{fit_bbox_to_aspect_ratio}
\alias{fit_bbox_to_aspect_ratio}
\title{Update the bbox to align with a specified aspect ratio}
\usage{
fit_bbox_to_aspect_ratio(bbox, target_aspect_ratio = 1.75)
}
\arguments{
\item{bbox}{Object of class bbox}

\item{target_aspect_ratio}{Aspect ratio to convert the bbox to, defaults to 7:4}
}
\value{
updated object of class bbox
}
\description{
Update the bbox to align with a specified aspect ratio
}
