% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\docType{data}
\name{cori_font}
\alias{cori_font}
\alias{tthoves_install}
\alias{tthoves_test}
\alias{load_fonts}
\title{Test, import, and register TT Hoves font}
\format{
An object of class \code{character} of length 1.
}
\usage{
cori_font

tthoves_install()

tthoves_test()

load_fonts(load_TT_Hoves = FALSE)
}
\arguments{
\item{load_TT_Hoves}{Boolean to determine if TT Hoves should be loaded from a local fontbook. Defaults to FALSE.}
}
\description{
You need to install the TT Hoves font locally to use the CORI theming.
}
\details{
\itemize{
\item \code{tthoves_test()} tests to see if the TT Hoves font is imported and
registered.
\item \code{tthoves_install()} registers the TT Hoves font for use in R.
}
}
\examples{
cori_font
try(tthoves_test())
}
\keyword{datasets}
