% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold_annotate.R
\name{geom_threshold_annotate}
\alias{geom_threshold_annotate}
\title{Add a vertical or horizontal line with an annotation}
\usage{
geom_threshold_annotate(intercept, axis = c("x", "y"), label = "", shift = 0)
}
\arguments{
\item{intercept}{x or y intercept (indicated by \code{axis}) through which the
vertical or horizontal threshold is made.}

\item{axis}{Axis along which a threshold is to be added, defaults to "x"
indicating a vertical threshold}

\item{label}{Label to add to the graph, defaults to an empty string}

\item{shift}{How much to shift the label in \code{shift} units above the
threshold line along "y", or in \code{shift} units to the right of the
threshold line along "x", defaults to 0}
}
\description{
This is a utility function that allows the user to add a line threshold with
text annotation using an intercept parameter.
The shift parameter allows the user to relatively adjust the text with
respect to the defined threshold value.
}
\examples{

library(ggplot2)

ggplot(mtcars, aes(wt, mpg)) +
  geom_point() +
  geom_threshold_annotate(20, axis = "y", shift = 1, label = "20 mpg")
}
