% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/watermark.R
\name{geom_watermark}
\alias{geom_watermark}
\title{Watermark}
\usage{
geom_watermark(alpha = 0.2)
}
\arguments{
\item{alpha}{Opacity for the watermark, defaults to 0.2}
}
\description{
Add the CORI watermark as a background to a \link{ggplot2} object
}
\examples{
library(ggplot2)
library(dplyr)
data(cori_employment)
cori_employment \%>\%
  ggplot(aes(estimate_pop_2014, estimate_pop_2019)) +
  geom_watermark(alpha = 0.1) +
  geom_point(size = 4, alpha = .6)

}
