% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_chart_text_from_gsheet}
\alias{get_chart_text_from_gsheet}
\title{Pull chart text from Google Sheet and format it}
\usage{
get_chart_text_from_gsheet(
  sheet_id,
  viz_unique_id,
  title_wrap = 150,
  subtitle_wrap = 150,
  caption_wrap = 80
)
}
\arguments{
\item{sheet_id}{Character. The Google Sheets ID.}

\item{viz_unique_id}{Character. The unique visualization ID in the sheet.}

\item{title_wrap}{Integer. Character width for wrapping the title. Default is 150.}

\item{subtitle_wrap}{Integer. Character width for wrapping the subtitle. Default is 150.}

\item{caption_wrap}{Integer. Character width for wrapping the source and notes. Default is 80.}
}
\value{
A named list with elements: \code{title}, \code{subtitle} (if available), \code{x} (if available), \code{y} (if available), and \code{caption} (if available).
}
\description{
This function retrieves chart-related text from a Google Sheet and formats it.
It assumes that the spreadsheet contains the following columns: \code{viz_id}, \code{title}, \code{subtitle}, \code{source}, \code{note}, \code{x}, and \code{y}.
}
