% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_cori}
\alias{scale_fill_cori}
\alias{scale_color_cori}
\title{Color scales for CORI data visualization}
\usage{
scale_fill_cori(palette = "ctg2gn", discrete = TRUE, reverse = FALSE, ...)

scale_color_cori(palette = "ctg2gn", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Character name of palette in \code{\link[=cori_palettes]{cori_palettes()}}}

\item{discrete}{Boolean indicating whether color aesthetic is discrete or not.
Defaults to \code{TRUE}}

\item{reverse}{Boolean indicating whether the palette should be reversed.
Defaults to \code{FALSE}}

\item{...}{Additional arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}} or
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradientn()}}, used respectively when \code{discrete} is
\code{TRUE} or \code{FALSE}}
}
\description{
Color scales for CORI data visualization
}
\examples{
library(ggplot2)
library(dplyr)
data(cori_employment)
cori_employment \%>\%
  ggplot(
    aes(estimate_pop_2014, estimate_pop_2019,
      color = estimate_employed_2019
    )
  ) +
  geom_point(size = 4, alpha = .6) +
  scale_color_cori(discrete = FALSE, palette = "ctg2pu")

library(tidyr)
library(stringr)
cori_employment \%>\%
  pivot_longer(contains("employed")) \%>\%
  mutate(name = str_remove(name, "estimate_employed_")) \%>\%
  ggplot(aes(value, rin_community, fill = name)) +
  geom_col(position = "dodge") +
  scale_fill_cori() +
  labs(x = "Estimated population employed", y = NULL, fill = NULL)
}
