% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_cori_map}
\alias{theme_cori_map}
\title{A \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} theme for the Center on Rural Innovation (CORI) style}
\usage{
theme_cori_map(title_family = "Lato", base_family = "Lato", base_size = 15)
}
\arguments{
\item{title_family}{Font family that CORI employs for titles, defaults to  "Lato"}

\item{base_family}{Font family that CORI employs, defaults to  "Lato"}

\item{base_size}{Base text font size, defaults to 12}
}
\description{
\code{theme_cori_map()} provides a \link[ggplot2:ggplot2-package]{ggplot2::ggplot2} theme formatted according to the CORI
style guide for maps (choropleth or dot density)
}
