% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{add_logo_svg}
\alias{add_logo_svg}
\title{Add CORI logo ggplot figure in an SVG-friendly way}
\usage{
add_logo_svg(
  fig,
  logo_path = "https://rwjf-public.s3.amazonaws.com/Logo-Mark_CORI_Black.svg",
  x_pos_scale = 0.9975,
  y_pos_scale = 1.195
)
}
\arguments{
\item{fig}{ggplot2 figure}

\item{logo_path}{Path to the logo. Defaults to hosted Full CORI Black logo}

\item{x_pos_scale}{Position scale factor as a percentage of x range}

\item{y_pos_scale}{Position scale factor as a percentage of the y range}
}
\value{
ggplot figure with logo
}
\description{
Add CORI logo ggplot figure in an SVG-friendly way
}
