% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{cori_colors}
\alias{cori_colors}
\alias{cori_cols}
\alias{cori_palettes}
\alias{cori_pal}
\title{Functions for handling CORI colors}
\format{
An object of class \code{character} of length 29.

An object of class \code{list} of length 12.
}
\usage{
cori_colors

cori_cols(...)

cori_palettes

cori_pal(palette = "ctg2gn", reverse = FALSE, ...)
}
\arguments{
\item{...}{Additional arguments to pass to \link[grDevices:colorRamp]{grDevices::colorRampPalette}}

\item{palette}{Character name of palette in \link{cori_palettes}}

\item{reverse}{Boolean indicating whether the palette should be reversed}
}
\description{
These utility functions provide fluent access to CORI colors and palettes in
formats ready for \link{ggplot2}.
}
\examples{
cori_colors["Emerald"]
cori_cols("Emerald")
cori_palettes[["ctg2gn"]]
cori_pal("ctg7")
}
\keyword{datasets}
