% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{load_fonts}
\alias{load_fonts}
\title{Functions for working with CORI fonts}
\usage{
load_fonts(fonts = c("Lato", "Bitter"))
}
\arguments{
\item{fonts}{Character vector of Google font names to load.}
}
\description{
Functions for working with CORI fonts
}
