% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fonts.R
\name{tthoves_install}
\alias{tthoves_install}
\alias{tthoves_test}
\title{Test, import, and register TT Hoves font}
\usage{
tthoves_install()

tthoves_test()
}
\description{
You need to install the TT Hoves font locally to use the CORI theming.
}
\details{
\itemize{
\item \code{tthoves_test()} tests to see if the TT Hoves font is imported and
registered.
\item \code{tthoves_install()} registers the TT Hoves font for use in R.
}
}
\examples{
try(tthoves_test())

}
