% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{save_plot}
\alias{save_plot}
\title{Save a plot with CORI defaults}
\usage{
save_plot(
  fig,
  export_path,
  chart_width = 640,
  chart_height = 450,
  add_logo = TRUE,
  logo_path = "https://rwjf-public.s3.amazonaws.com/Logo-Mark_CORI_Black.svg",
  logo_position = "top right",
  logo_scale = 20,
  background = "white"
)
}
\arguments{
\item{fig}{The ggplot2 figure}

\item{export_path}{file path for the exported plot}

\item{chart_width}{The width in px of the chart}

\item{chart_height}{The height in px of the chart}

\item{add_logo}{Boolean that determines if a logo is included or not}

\item{logo_path}{Path to the logo. Defaults to hosted Full CORI Black logo}

\item{logo_position}{Combination of top/bottom and right/left. Defaults to top right.}

\item{logo_scale}{Scale logo to 1/10 width of plot}

\item{background}{Color of the background in the image export}
}
\description{
Save a plot with CORI defaults
}
