% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme.R
\name{theme_cori_map_presentation}
\alias{theme_cori_map_presentation}
\title{A \link{ggplot2} theme for the Center on Rural Innovation (CORI) style}
\usage{
theme_cori_map_presentation(base_family = "Lato", base_size = 12)
}
\arguments{
\item{base_family}{Font family that CORI employs, defaults to  "Lato"}

\item{base_size}{Base text font size, defaults to 12}
}
\description{
\code{theme_cori_map_presentation()} provides a \link{ggplot2} theme formatted for maps
used in presentations
}
