% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_defaults.R
\name{update_cori_geom_defaults}
\alias{update_cori_geom_defaults}
\title{The Center for Rural Innovation \link{ggplot2} theme}
\usage{
update_cori_geom_defaults(base_family = "Montserrat")
}
\arguments{
\item{base_family}{The base font family for the theme, defaults to "Montserrat".}
}
\description{
\code{update_cori_geom_defaults()} provides a \link{ggplot2} theme that aligns with the
CORI style guide, with sensible defaults.
}
\examples{
library(ggplot2)
library(dplyr)
data(cori_employment)

update_cori_geom_defaults()
cori_education \%>\%
  group_by(date) \%>\%
  summarize(percent_working_remotely = mean(percent_working_remotely)) \%>\%
  ggplot(aes(date, percent_working_remotely)) +
  geom_hline(yintercept = 0.12) +
  scale_y_continuous(labels = scales::percent) +
  geom_line() +
  geom_point()
}
